Move objects-api logic into Vuex store module
