+++ src/store/index.js
@@ -1,11 +1,12 @@
 import Vue from "vue"
 import Vuex from "vuex"
 import selected from "./modules/selected"
+import objects from "./modules/objects"
 
 Vue.use(Vuex)
 
 export default new Vuex.Store({
   modules: {
-    selected
+    selected, objects
   },
 })
